+++ o3seespy/__init__.py
@@ -12,6 +12,7 @@ from .command import test_check  # deprecated
 from .__about__ import __version__
 from . import results
 from . import mptools, mp
+from .command import parallelcmds
 
 try:
     from custom_openseespy import opensees as opy